+++ server/middlewares/authMiddleware.js
@@ -1,5 +1,5 @@
 import jwt from 'jsonwebtoken';
-import User from "../models/User.js";
+import User from "../models/UserModel.js";
 
 const secretKey = process.env.JWT_SECRET_KEY;
