+++ server/models/IncomeModel.js
@@ -1,6 +1,6 @@
 import {DataTypes} from "sequelize";
 import sequelize from "../config/database.js";
-import User from "./User.js";
+import User from "./UserModel.js";
 
 const Income = sequelize.define('Income', {
   date: {
